Rename auth and auth provider for readability
